modified:   enable_internet-detector.sh

+++ enable_internet-detector.sh
@@ -3,31 +3,33 @@
 
 set -x
 
-# exit if routerich #
 vBOARD_ID="$(cat /etc/board.json | grep id | sed 's/\"/ /g;s/,/ /g' | awk '{print $3}')"
-[ "${vBOARD_ID}" = "routerich" ] && (echo " --- DETECTED BOARD: ${vBOARD_ID} --- Exit with code '1' ---"; exit 1)
 
-vNAME="internet-detector"
-vVERSION="1.4.2-r1"
-vFILE="${vNAME}_${vVERSION}_all.ipk"
-vFILELUCI="luci-app-${vNAME}_${vVERSION}_all.ipk"
-vFILELUCILANG="luci-i18n-${vNAME}-ru_${vVERSION}_all.ipk"
-vURL="https://github.com/gSpotx2f/packages-openwrt/raw/master/current"
+if [ "${vBOARD_ID}" == "routerich" ]; then 
+    echo " --- DETECTED BOARD: routerich ---"
+else
+    vNAME="internet-detector"
+    vVERSION="1.4.2-r1"
+    vFILE="${vNAME}_${vVERSION}_all.ipk"
+    vFILELUCI="luci-app-${vNAME}_${vVERSION}_all.ipk"
+    vFILELUCILANG="luci-i18n-${vNAME}-ru_${vVERSION}_all.ipk"
+    vURL="https://github.com/gSpotx2f/packages-openwrt/raw/master/current"
 
-wget ${vURL}/${vFILE} -O /tmp/${vFILE}
-wget ${vURL}/${vFILELUCI} -O /tmp/${vFILELUCI}
-wget ${vURL}/${vFILELUCILANG} -O /tmp/${vFILELUCILANG}
+    wget ${vURL}/${vFILE} -O /tmp/${vFILE}
+    wget ${vURL}/${vFILELUCI} -O /tmp/${vFILELUCI}
+    wget ${vURL}/${vFILELUCILANG} -O /tmp/${vFILELUCILANG}
 
-opkg install /tmp/${vFILE}
-/etc/init.d/internet-detector enable
-/etc/init.d/internet-detector start
+    opkg install /tmp/${vFILE}
+    /etc/init.d/internet-detector enable
+    /etc/init.d/internet-detector start
 
-opkg install /tmp/${vFILELUCI}
-/etc/init.d/rpcd restart
+    opkg install /tmp/${vFILELUCI}
+    /etc/init.d/rpcd restart
 
-opkg install /tmp/${vFILELUCILANG}
+    opkg install /tmp/${vFILELUCILANG}
 
-rm -fv /tmp/*${vNAME}*
+    rm -fv /tmp/*${vNAME}*
+fi
 
 set +x
